backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.4.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.3
